app/requirements.txt
datasets==3.2.0
fastapi==0.115.8
haystack-ai==2.9.0
mistral-haystack
python-dotenv
sentence-transformers==3.4.1
transformers==4.48.2
pydantic==2.10.6
pydantic-settings==2.7.1
uvicorn==0.34.0
